+++ src/main/kotlin/com/github/jensim/megamanipulator/listener/OnStartListener.kt
@@ -1,0 +1,23 @@
+package com.github.jensim.megamanipulator.listener
+
+import com.github.jensim.megamanipulator.onboarding.OnboardingOperator
+import com.github.jensim.megamanipulator.project.MegaManipulatorUtil.isMM
+import com.intellij.openapi.project.Project
+import com.intellij.openapi.startup.StartupActivity
+import com.intellij.openapi.wm.ToolWindowManager
+
+class OnStartListener : StartupActivity {
+
+    override fun runActivity(project: Project) {
+        if (isMM(project)) {
+            OnboardingOperator.registerProject(project)
+            ToolWindowManager.getInstance(project)?.let {
+                it.getToolWindow("Project")?.show()
+            }
+            ToolWindowManager.getInstance(project)?.let {
+                it.getToolWindow("Mega Manipulator")?.show()
+            }
+            // TODO OnboardingOperator.display(OnboardingId.WELCOME)
+        }
+    }
+}