+++ src/main/kotlin/com/github/jensim/megamanipulator/onboarding/OnboardingId.kt
@@ -1,0 +1,16 @@
+package com.github.jensim.megamanipulator.onboarding
+
+enum class OnboardingId(val title: String, val text: String) {
+    CONFIG_FILE(
+        "Config #️⃣",
+        "This config is where you define your code and search hosts",
+    ),
+    SETTINGS_TAB(
+        "Settings ⚙️",
+        "Here you are able to validate the settings files, set and validate your login credentials to the code & search hosts",
+    ),
+    WELCOME(
+        "Welcome! ☀️",
+        "This is where you find the mega-manipulator",
+    ),
+}