+++ src/test/kotlin/com/github/jensim/megamanipulator/actions/vcs/bitbucketserver/BitbucketServerClientTest.kt
@@ -3,7 +3,6 @@ package com.github.jensim.megamanipulator.actions.vcs.bitbucketserver
 import com.github.jensim.megamanipulator.actions.NotificationsOperator
 import com.github.jensim.megamanipulator.actions.localrepo.LocalRepoOperator
 import com.github.jensim.megamanipulator.http.HttpClientProvider
-import com.github.jensim.megamanipulator.settings.SerializationHolder
 import com.github.jensim.megamanipulator.settings.SettingsFileOperator
 import com.github.jensim.megamanipulator.settings.types.CloneType.HTTPS
 import com.github.jensim.megamanipulator.settings.types.CodeHostSettings.BitBucketSettings
@@ -69,7 +68,6 @@ internal class BitbucketServerClientTest {
     private val client = BitbucketServerClient(
         httpClientProvider = clientProvider,
         localRepoOperator = localRepoMock,
-        json = SerializationHolder.instance.readableJson
     )
 
     @Test