+++ src/test/kotlin/com/github/jensim/megamanipulator/settings/SettingsFileOperatorTest.kt
@@ -1,5 +1,6 @@
 package com.github.jensim.megamanipulator.settings
 
+import com.github.jensim.megamanipulator.settings.SerializationHolder.readableJson
 import com.github.jensim.megamanipulator.settings.types.CodeHostSettings
 import com.github.jensim.megamanipulator.settings.types.MegaManipulatorSettings
 import com.github.jensim.megamanipulator.settings.types.SearchHostSettings
@@ -7,7 +8,6 @@ import com.github.ricky12awesome.jss.encodeToSchema
 import com.github.ricky12awesome.jss.globalJson
 import kotlinx.serialization.decodeFromString
 import kotlinx.serialization.encodeToString
-import kotlinx.serialization.json.Json
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Disabled
@@ -16,7 +16,6 @@ import java.io.File
 
 class SettingsFileOperatorTest {
 
-    private val readableJson: Json = SerializationHolder.instance.readableJson
     private val testData = MegaManipulatorSettings(
         searchHostSettings = mapOf(
             "sourcegraph_com" to SearchHostSettings.SourceGraphSettings(